Final Paquetes Admin Daniel Carrillo

+++ orderApp/src/main/java/utez/edu/mx/orderapp/controllers/packages/dtos/PackageInfoDto.java
@@ -16,7 +16,7 @@ public class PackageInfoDto {
     private Long packageId;
     private String packageName;
     private String packageDescription;
-    private Float packagePrice;
+    private Long packagePrice;
     private Boolean packageState;
     private Integer designatedHours;
     private Integer workersNumber;
